projects/lesson3_chains_primal/requirements.txt
langchain
langchain-openai
langchain-community
openai
python-dotenv
pydantic
colorama